Fix claim stars double vs areas size int issue

+++ src/main/java/me/nurio/imperial/core/areas/PlayerClaimListener.java
@@ -61,7 +61,7 @@ public class PlayerClaimListener implements Listener {
         // Not enough power
         double stars = organization.getStars();
         int currentSize = organization.getWorldArea().getAreas().size();
-        if (stars <= currentSize) {
+        if (stars <= (double) currentSize) {
             Bukkit.getLogger().warning("Tried to claim but not enough stars");
             return;
         }
